rename: 프로필 관련 API 요청 함수의 폴더 구조 변경
- 프로필 관련 API 요청 함수를 profile 폴더 내로 옮겼습니다.
