+++ src/pages/Profile/index.tsx
@@ -6,7 +6,7 @@ import Header from './components/Header'
 import PostList from './components/profile/PostList'
 import Drawer from './components/profile-edit-drawer/Drawer'
 
-import getProfile from '@/apis/profile'
+import getProfile from '@/apis/profile/profile'
 import { useProfileStore } from '@/stores/userProfileStore'
 import ProfileImage from './components/profile-edit-drawer/ProfileImage'
 import CoverImage from './components/profile-edit-drawer/CoverImage'
